Completely rewrote how data is passed between protocols
Protocols used to have two layers of indirection: the first was the
protocol itself, which served mostly as a workflow definition, and the
second was the scripts responsible for the actual work. This was really
cumbersome, so I rewrote it all. Now, protocols directly implement what
needs to get done. This means there aren't really "classes" of protocols
anymore, but it also means that there's a lot less cruft. Most
importantly, data exists natively in protocols, as subclasses of
Bundles. This streamlines the production and consumption of data, as it
can be produced and stored in its native form without the need for extra
JSON files like we had before.

NOTE: THIS IS AN INCOMPATIBLE CHANGE. Protocols written for protos
before this commit will not function after it.