Minor fixes and temporarily disabled 'add_files'.
Fixed a minor timezone bug in data_bundles.

Removed the functionality of add_files because it was assuming the old
directory-based bundle structure. Now that we have abstracted storage
from bundles, we need a new method. This probably means an interface
change to storage and an implementation for each storage mechanism.

Also added a void_bundle for protocols that don't return anything.