Added checks for NULL tags in Postgres adapter.
We were bombing out a bunch on un-specified tags fields, since we special-case
it and apply a json (de-)serialization to it. Normally, it should be a '[]'
string if it no tags were given. However, that's only if protos created it
from an Experiment instance. For raw entries (obsolete rows and test code), it
could be left as an SQL NULL. This caused the json parsers to die. Now we check.