Allow protos to function without all storage_adapters.
Storage adapters are the most common missing support package for protos.
Moreover, sometimes we don't even need all of them. Instead of forcing the
user to install unnecessary packages, we allow protos to start with missing
storage adapters disabled. Obviously, if the user tries to use them, protos
will complain about a missing adapter, but otherwise it should function fine.

The current behavior is to warn on a failed storage adapter.